+++ extensions/teleport/src/list-databases.tsx
@@ -1,5 +1,5 @@
 import { List, ActionPanel, Action, showToast, Toast, getPreferenceValues, Form } from "@raycast/api";
-import { usePromise } from "@raycast/utils";
+import { useCachedPromise } from "@raycast/utils";
 import { getDatabasesList, connectToDatabase } from "./utils";
 
 async function open(name: string, protocol: string, database: string) {
@@ -39,7 +39,7 @@ function DatabaseForm(props: { name: string; protocol: string }) {
 }
 
 export default function Command() {
-  const { data, isLoading } = usePromise(getDatabasesList);
+  const { data, isLoading } = useCachedPromise(getDatabasesList);
 
   return (
     <List isLoading={isLoading}>