+++ extensions/teleport/src/list-servers.tsx
@@ -1,6 +1,6 @@
 import { List, ActionPanel, Action, showToast, Toast, getPreferenceValues } from "@raycast/api";
 import { runAppleScript } from "run-applescript";
-import { usePromise } from "@raycast/utils";
+import { useCachedPromise } from "@raycast/utils";
 import { getServersList, getServerCommand } from "./utils";
 
 async function open(name: string) {
@@ -35,7 +35,7 @@ async function open(name: string) {
 }
 
 export default function Command() {
-  const { data, isLoading } = usePromise(getServersList);
+  const { data, isLoading } = useCachedPromise(getServersList);
 
   return (
     <List isLoading={isLoading}>