+++ extensions/teleport/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Teleport Changelog
 
+## [New Commands] - 2023-11-05
+- Add support for interacting with Kubernetes clusters
 
 ## [New Commands] - 2023-11-03
 - Add support for searching applications